+++ README.md
@@ -3,8 +3,7 @@
 ![Build status](https://github.com/dataiku/dss-plugin-nlp-analysis/actions/workflows/auto-make.yml/badge.svg) ![GitHub release (latest by date)](https://img.shields.io/github/v/release/dataiku/dss-plugin-nlp-analysis?logo=github)  ![Support level](https://img.shields.io/badge/support-Unsupported-orange)
 
 This Dataiku DSS plugin provides recipes to analyze text data:
-- TODO
-
+TODO
 Documentation: TODO
 
 ## License
