+++ custom-recipes/nlp-analysis-ontology-tagging/recipe.json
@@ -17,7 +17,7 @@
         {
             "name": "ontology_dataset",
             "label": "Ontology dataset",
-            "description": "Dataset with columns for tags, keywords and categories",
+            "description": "Dataset with one column for keywords (set of terms to match with text), tags (term assigned to a set of keywords) and categories(term assigned to a set of tags). Example: tag = 'New York', keyword = 'N.Y', category = 'city'",
             "arity": "UNARY",
             "required": true,
             "acceptsDataset": true
@@ -27,7 +27,7 @@
         {
             "name": "tagged_documents",
             "label": "Tagged documents",
-            "description": "Dataset with the input documents and their assigned tags",
+            "description": "Input dataset enriched with tags and keywords, and extracted matched sentences",
             "arity": "UNARY",
             "required": true,
             "acceptsDataset": true
@@ -392,16 +392,16 @@
             "type": "SELECT",
             "selectChoices": [
                 {
-                    "value": "one_row_per_tag",
-                    "label": "1 row / tag"
+                    "value": "one_row_per_match",
+                    "label": "row per match"
                 },
                 {
                     "value": "one_row_per_doc",
-                    "label": "1 row / doc (array columns)"
+                    "label": "row per doc (array columns)"
                 },
                 {
                     "value": "one_row_per_doc_json",
-                    "label": "1 row / doc (nested JSON)"
+                    "label": "row per doc (nested JSON)"
                 }
             ],
             "defaultValue": "one_row_per_doc"
