+++ python-lib/ontology_tagging/dku_plugin_config_loading.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-import dataiku
+from dataiku import Dataset
 from dataiku.customrecipe import (
     get_recipe_config,
     get_input_names_for_role,
@@ -29,11 +29,11 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
         super().__init__()
         text_input = get_input_names_for_role("document_dataset")[0]
         self.dku_config.add_param(
-            name="text_input", value=dataiku.Dataset(text_input), required=True
+            name="text_input", value=Dataset(text_input), required=True
         )
         ontology_input = get_input_names_for_role("ontology_dataset")[0]
         self.dku_config.add_param(
-            name="ontology_input", value=dataiku.Dataset(ontology_input), required=True
+            name="ontology_input", value=Dataset(ontology_input), required=True
         )
         self.document_dataset_columns = [
             p["name"] for p in self.dku_config.text_input.read_schema()
@@ -41,6 +41,26 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
         self.ontology_dataset_columns = [
             p["name"] for p in self.dku_config.ontology_input.read_schema()
         ]
+
+    def load_settings(self):
+        """Public function to load all given parameters for Ontology Tagging Plugin"""
+
+        self._add_matching_settings()
+        self._add_text_column()
+        self._add_language()
+        if self.dku_config.language == "language_column":
+            self._add_language_column()
+            self._add_lemmatization(multilingual=True)
+        else:
+            self.dku_config.add_param(
+                name="language_column",
+                value=None,
+            )
+            self._add_lemmatization()
+        self._add_ontology_columns()
+        self._add_output_format()
+        self._add_output_dataset()
+        return self.dku_config
 
     def _content_error_message(self, error, column):
         """Get corresponding error message if any"""
@@ -125,22 +145,13 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
                 },
                 {
                     "type": "custom",
-                    "cond": lang_column in input_columns + [None]
-                    or not bool(lang_column),
+                    "cond": lang_column in input_columns,
                     "err_msg": self._content_error_message("invalid", lang_column),
                 },
             ],
         )
 
-    def _check_languages(self, languages):
-        """Checks if the specified languages are supported"""
-        unsupported_languages = set(languages) - set(SUPPORTED_LANGUAGES_SPACY.keys())
-        if unsupported_languages:
-            raise ValueError(
-                f"Found {len(unsupported_languages)} unsupported languages in Document dataset: {unsupported_languages}"
-            )
-
-    def _check_languages(self, languages):
+    def check_languages(self, languages):
         """Checks if the specified languages are supported"""
         unsupported_languages = set(languages) - set(SUPPORTED_LANGUAGES_SPACY.keys())
         if unsupported_languages:
@@ -158,12 +169,12 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
             },
             {
                 "type": "custom",
-                "cond": column in input_columns or column == None,
+                "cond": column in input_columns,
                 "err_msg": self._content_error_message("invalid", column),
             },
         ]
 
-    def _ontology_columns_mandatory(self, column_name, column_label, input_columns):
+    def _ontology_columns_mandatory(self, column_name, input_columns):
         """Load mandatory columns from Ontology Dataset"""
 
         column = self.config.get(column_name)
@@ -178,10 +189,8 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
         """Load columns from Ontology Dataset"""
 
         input_columns = self.ontology_dataset_columns
-        self._ontology_columns_mandatory("tag_column", "Tag column", input_columns)
-        self._ontology_columns_mandatory(
-            "keyword_column", "Keyword column", input_columns
-        )
+        self._ontology_columns_mandatory("tag_column", input_columns)
+        self._ontology_columns_mandatory("keyword_column", input_columns)
         ontology_columns = [self.dku_config.tag_column, self.dku_config.keyword_column]
         self._add_category_column()
         category_column = self.dku_config.category_column
@@ -221,26 +230,6 @@ class DkuConfigLoadingOntologyTagging(DkuConfigLoading):
         output_dataset_name = get_output_names_for_role("tagged_documents")[0]
         self.dku_config.add_param(
             name="output_dataset",
-            value=dataiku.Dataset(output_dataset_name),
+            value=Dataset(output_dataset_name),
             required=True,
         )
-
-    def load_settings(self):
-        """Public function to load all given parameters for Ontology Tagging Plugin"""
-
-        self._add_matching_settings()
-        self._add_text_column()
-        self._add_language()
-        if self.dku_config.language == "language_column":
-            self._add_language_column()
-            self._add_lemmatization(multilingual=True)
-        else:
-            self.dku_config.add_param(
-                name="language_column",
-                value="",
-            )
-            self._add_lemmatization()
-        self._add_ontology_columns()
-        self._add_output_format()
-        self._add_output_dataset()
-        return self.dku_config
